LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-selectmenu-element/selectmenu-parts-structure.tentative-expected.txt

three
Custom button
Custom button
This is some text. More text.
button0
button0
button0
Test
button0
button0
button1
button2
button0


FAIL HTMLOptionElements (and not other element types) should receive option controller code assert_equals: expected (string) "one" but got (undefined) undefined
FAIL To receive option part controller code, an option must be a descendant of the listbox part in a flat tree traversal assert_equals: expected (string) "one" but got (undefined) undefined
FAIL To receive button part controller code, an element labeled as a button must not be a descendant of the listbox part in a flat tree traversal promise_test: Unhandled rejection with value: object "SyntaxError: ':top-layer' is not a valid selector."
FAIL To receive listbox part controller code, an element labeled as a listbox must not be a descendant of the button part in a flat tree traversal promise_test: Unhandled rejection with value: object "SyntaxError: ':top-layer' is not a valid selector."
FAIL Button controller code should be applied in flat tree traversal order regardless of dynamic insertion order assert_false: expected false got undefined
FAIL Listbox controller code should be applied in flat tree traversal order regardless of dynamic insertion order assert_false: expected false got undefined
FAIL selected-value controller code should be applied in flat tree traversal order regardless of dynamic insertion order assert_false: expected false got undefined
FAIL Ensure that option controller code is updated when listbox changes assert_equals: Initial value should be the first option expected (string) "one" but got (undefined) undefined
FAIL Ensure that controller code is applied after updating the slot attribute assert_false: Controller code not applied due to part attribute missing expected false got undefined
FAIL Ensure that controller code is applied when slot and part attributes are different assert_false: Controller code not applied due to part attribute not being button expected false got undefined
FAIL Ensure that controller code is synchronously applied assert_equals: Initial value should be the first option expected (string) "one" but got (undefined) undefined
FAIL Controller code should be updated when nested parts are removed assert_equals: expected (string) "one" but got (undefined) undefined
FAIL Moving a selectmenu between documents should keep controller code active assert_equals: expected (string) "one" but got (undefined) undefined
FAIL New parts should be detected even when in the subtree of an inserted node assert_true: New button should receive controller code expected true got undefined
FAIL Part removals should be detected even when in the subtree of a removed node assert_false: Removed button should no longer have controller code expected false got undefined
